edit

+++ File Management/src/edu/najah/cap/FileRepository/impl/NormalDelete.java
@@ -2,12 +2,16 @@ package edu.najah.cap.FileRepository.impl;
 
 import edu.najah.cap.Database.impl.MySQLDatabase;
 import edu.najah.cap.FileRepository.intf.Delete;
+import edu.najah.cap.Security.Authorization;
 
 public class NormalDelete implements Delete {
 
     @Override
     public void delete(String filename, String category) {
         try {
+            if(!Authorization.hasAdminPermission()){
+                return;
+            }
             String query = "DELETE FROM files WHERE name = "+filename+" AND category = "+category+";";
             MySQLDatabase.getInstance().execute(query);
         }catch (Exception e){
